+++ gldcore/test.cpp
@@ -105,7 +105,7 @@ STATUS test_exec(void)
 static volatile unsigned int *count = NULL;
 static volatile unsigned int total = 0;
 static LOCKVAR key = 0;
-static volatile int done = 0;
+static volatile size_t done = 0;
 
 static void *test_lock_proc(void *ptr)
 {
@@ -128,7 +128,7 @@ static void *test_lock_proc(void *ptr)
 
 int test_lock(void)
 {
-	int n, sum=0;
+	size_t n, sum=0;
 
 	count = (unsigned int*)malloc(sizeof(unsigned int*)*global_threadcount);
 	if ( !count )
@@ -139,7 +139,7 @@ int test_lock(void)
 	
 	output_test("*** Begin memory locking test for %d threads", global_threadcount);
 	wlock(&key);
-	for ( n=0 ; n<global_threadcount ; n++ )
+	for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 	{
 		pthread_t pt;
 		count[n] = 0;
@@ -151,28 +151,28 @@ int test_lock(void)
 	}
 	wunlock(&key);
 	global_suppress_repeat_messages = 0;
-	for ( n=0 ; n<global_threadcount ; n++ )
+	for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 		output_raw("THREAD %2d  ", n);
 	output_message("  TOTAL     ERRORS");
-	for ( n=0 ; n<global_threadcount ; n++ )
+	for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 		output_raw("---------- ", n);
 	output_message("---------- --------");
-	while ( done<global_threadcount )
+	while ( done < (size_t)global_threadcount )
 	{
 		int c[256], t, s=0;
 		exec_sleep(100000);
 		output_raw("\r");
 		rlock(&key);
-		for ( n=0 ; n<global_threadcount ; n++ )
+		for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 			s += (c[n]=count[n]);
 		t = total;
 		runlock(&key);
-		for ( n=0 ; n<global_threadcount ; n++ )
+		for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 			output_raw("%10d ",c[n]);
 		output_raw("%10d %8d",t,t-s);
 	}
 	output_message("");
-	for ( n=0 ; n<global_threadcount ; n++ )
+	for ( n=0 ; n < (size_t)global_threadcount ; n++ )
 	{
 		output_test("thread %d count = %d", n, count[n]);
 		sum+=count[n];
